DownloadMetrics: refactor
