+++ internal/firebase/structs/structs.go
@@ -56,3 +56,15 @@ type IntegerValue struct {
 type StringValue struct {
 	StringValue string `json:"stringValue"`
 }
+
+//MetricsData Data of metrics.
+type MetricsData struct {
+	Modified               int64  `json:"modified"`
+	Date                   string `json:"date"`
+	ActivationsYesterday   int32  `json:"activations_yesterday"`
+	ActivationsTotal       int32  `json:"activations_total"`
+	KeyPublishersYesterday int32  `json:"key_publishers_yesterday"`
+	KeyPublishersTotal     int32  `json:"key_publishers_total"`
+	NotificationsYesterday int32  `json:"notifications_yesterday"`
+	NotificationsTotal     int32  `json:"notifications_total"`
+}
